[WIP] 2019 Q1 OKR's (#379)
I'm trying something a little different with the Q1 OKR's this time around. Instead of a spreadsheet I'm going to see if we can use a markdown file in the repo.

This is an experiment, it may succeed or it may fail horribly :) It's certainly **not** easier to write OKR's in this format, but it does add something to the discoverability and once I put a link to an active issue in here for every OKR it'll be a lot more useful.